Update dependency versions

+++ src/transport/MessageEndpoint.cpp
@@ -123,7 +123,7 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                 }
                 default: {
                     SPDLOG_ERROR(
-                      "Invalid bind socket type {} ({})", socketType, address);
+                      "Invalid bind socket type {} ({})", static_cast<int>(socketType), address);
                     throw std::runtime_error(
                       "Binding with invalid socket type");
                 }
@@ -176,7 +176,7 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                 }
                 default: {
                     SPDLOG_ERROR("Invalid connect socket type {} ({})",
-                                 socketType,
+                                 static_cast<int>(socketType),
                                  address);
                     throw std::runtime_error(
                       "Connecting with unrecognized socket type");
